requirements.txt
slack-bolt>=1.18.0,<2
slack-sdk>=3.23.0,<4
tiktoken>=0.5.1,<0.6
# https://github.com/Yelp/elastalert/issues/2306
urllib3<2
anthropic
python-dotenv